Fixing OneLoginService/OneLoginUserService tests
